refactor: Introduce AppContent wrapper
- move integration of `jwtDecode`, `Routes`, `Route`, `Navigation`, `useLocation`, `socketEventManager`, `setUser` action and components from `App.js` to `appContent.js`.
- move socket instance initialization, and jwt decoding `useEffect` logic and snippets from `App.js` to `appContent.js`.
- add logic for conditionally rendering `Header` component based on the location name and token in `appContent.js`.
